new initial schema
