--- backend/alembic/versions/009efbef27ab_added_timezone.py
@@ -1,75 +1,0 @@
-"""added timezone
-
-Revision ID: 009efbef27ab
-Revises: 0deff22eb066
-Create Date: 2025-06-23 18:37:23.554175
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '009efbef27ab'
-down_revision: Union[str, None] = '0deff22eb066'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('articles', 'published_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.TIMESTAMP(timezone=True),
-               existing_nullable=False)
-    op.alter_column('articles', 'crawled_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.TIMESTAMP(timezone=True),
-               existing_nullable=False)
-    op.alter_column('articles', 'scraped_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.TIMESTAMP(timezone=True),
-               existing_nullable=True)
-    op.add_column('emails', sa.Column('update_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False))
-    op.alter_column('emails', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.TIMESTAMP(timezone=True),
-               existing_nullable=False)
-    op.drop_column('emails', 'updated_at')
-    op.alter_column('reports', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               type_=sa.TIMESTAMP(timezone=True),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('reports', 'created_at',
-               existing_type=sa.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False)
-    op.add_column('emails', sa.Column('updated_at', postgresql.TIMESTAMP(), autoincrement=False, nullable=False))
-    op.alter_column('emails', 'created_at',
-               existing_type=sa.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False)
-    op.drop_column('emails', 'update_at')
-    op.alter_column('articles', 'scraped_at',
-               existing_type=sa.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=True)
-    op.alter_column('articles', 'crawled_at',
-               existing_type=sa.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False)
-    op.alter_column('articles', 'published_at',
-               existing_type=sa.TIMESTAMP(timezone=True),
-               type_=postgresql.TIMESTAMP(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
