--- backend/alembic/versions/04bf14d0cb1f_author_to_json_format.py
@@ -1,68 +1,0 @@
-"""author to json format
-
-Revision ID: 04bf14d0cb1f
-Revises: 93ba3c1cbed0
-Create Date: 2025-06-16 22:59:28.431328
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = "04bf14d0cb1f"
-down_revision: Union[str, None] = "93ba3c1cbed0"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "articles",
-        sa.Column(
-            "authors",
-            sa.JSON(),
-            nullable=True,
-            comment="List of authors for the article",
-        ),
-    )
-    op.add_column(
-        "articles",
-        sa.Column(
-            "categories",
-            sa.JSON(),
-            nullable=True,
-            comment="List of categories for the article",
-        ),
-    )
-    op.drop_column("articles", "author")
-    op.drop_column("articles", "category")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "articles",
-        sa.Column(
-            "category",
-            postgresql.JSON(astext_type=sa.Text()),
-            autoincrement=False,
-            nullable=True,
-            comment="List of categories for the article",
-        ),
-    )
-    op.add_column(
-        "articles",
-        sa.Column("author", sa.VARCHAR(), autoincrement=False, nullable=True),
-    )
-    op.drop_column("articles", "categories")
-    op.drop_column("articles", "authors")
-    # ### end Alembic commands ###
