--- backend/alembic/versions/0afdeaced3f3_add_language_field_to_reports.py
@@ -1,41 +1,0 @@
-"""Add language field to reports
-
-Revision ID: 0afdeaced3f3
-Revises: 3e8a9b88e46c
-Create Date: 2025-06-26 13:20:49.811520
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = '0afdeaced3f3'
-down_revision: Union[str, None] = '3e8a9b88e46c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        'reports',
-        sa.Column(
-            'language',
-            sqlmodel.sql.sqltypes.AutoString(length=255),
-            nullable=False,
-            server_default="en"
-        )
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('reports', 'language')
-    # ### end Alembic commands ###
