--- backend/alembic/versions/4457e8b43ae3_added_crawler_stats.py
@@ -1,74 +1,0 @@
-"""added crawler stats
-
-Revision ID: 4457e8b43ae3
-Revises: 37080347559f
-Create Date: 2025-06-27 08:38:22.814317
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy import func
-import uuid
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = "4457e8b43ae3"
-down_revision: Union[str, None] = "37080347559f"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "crawl_stats",
-        sa.Column(
-            "id",
-            sa.Uuid(),
-            nullable=False,
-            server_default=sa.text("gen_random_uuid()"),
-        ),
-        sa.Column("subscription_id", sa.Uuid(), nullable=False),
-        sa.Column("total_successful", sa.Integer(), nullable=False),
-        sa.Column("total_attempted", sa.Integer(), nullable=False),
-        sa.Column(
-            "crawl_date",
-            sa.Date(),
-            nullable=False,
-            server_default=sa.text("CURRENT_DATE"),
-        ),
-        sa.Column("notes", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["subscription_id"],
-            ["subscriptions.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_crawl_stats_subscription_id"),
-        "crawl_stats",
-        ["subscription_id"],
-        unique=False,
-    )
-    op.add_column(
-        "articles",
-        sa.Column("note", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("articles", "note")
-    op.drop_index(
-        op.f("ix_crawl_stats_subscription_id"), table_name="crawl_stats"
-    )
-    op.drop_table("crawl_stats")
-    # ### end Alembic commands ###
