--- backend/alembic/versions/5a15fbb052d2_clean_up_email_table.py
@@ -1,35 +1,0 @@
-"""clean up email table
-
-Revision ID: 5a15fbb052d2
-Revises: d8c46ac01508
-Create Date: 2025-06-20 17:31:50.643029
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5a15fbb052d2'
-down_revision: Union[str, None] = 'd8c46ac01508'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('emails', 'attachment')
-    op.drop_column('emails', 'attachment_name')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('emails', sa.Column('attachment_name', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column('emails', sa.Column('attachment', sa.VARCHAR(), autoincrement=False, nullable=True))
-    # ### end Alembic commands ###
