--- backend/alembic/versions/61d6eb52c29c_added_email_columns.py
@@ -1,35 +1,0 @@
-"""added email columns
-
-Revision ID: 61d6eb52c29c
-Revises: 35741af0d46a
-Create Date: 2025-06-11 15:05:03.796815
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = '61d6eb52c29c'
-down_revision: Union[str, None] = '35741af0d46a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('search_profiles', sa.Column('organization_emails', sa.ARRAY(sa.String()), nullable=True))
-    op.add_column('search_profiles', sa.Column('profile_emails', sa.ARRAY(sa.String()), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('search_profiles', 'profile_emails')
-    op.drop_column('search_profiles', 'organization_emails')
-    # ### end Alembic commands ###
