--- backend/alembic/versions/82da1701bd5f_email_table.py
@@ -1,48 +1,0 @@
-"""email table
-
-Revision ID: 82da1701bd5f
-Revises: 61d6eb52c29c
-Create Date: 2025-06-12 17:02:43.614756
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = '82da1701bd5f'
-down_revision: Union[str, None] = '61d6eb52c29c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('emails',
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('sender', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('recipient', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('subject', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('content', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('content_type', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.Column('attempts', sa.Integer(), nullable=False),
-    sa.Column('attachment', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    sa.Column('attachment_name', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    sa.Column('state', sa.Enum('PENDING', 'SENT', 'RETRY', 'FAILED', name='emailstate'), nullable=False),
-    sa.Column('errors', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('emails')
-    # ### end Alembic commands ###
