--- backend/alembic/versions/92673c837211_add_email_chatbot_tables.py
@@ -1,78 +1,0 @@
-"""add_email_chatbot_tables
-
-Revision ID: 92673c837211
-Revises: f4218eba1395
-Create Date: 2025-07-03 13:40:32.240047
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "92673c837211"
-down_revision: Union[str, None] = "f4218eba1395"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "email_conversations",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "subject",
-            sqlmodel.sql.sqltypes.AutoString(length=500),
-            nullable=False,
-        ),
-        sa.Column(
-            "user_email",
-            sqlmodel.sql.sqltypes.AutoString(length=255),
-            nullable=False,
-        ),
-        sa.Column("created_at", sa.TIMESTAMP(timezone=True), nullable=False),
-        sa.Column("updated_at", sa.TIMESTAMP(timezone=True), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_email_conversations_user_email"),
-        "email_conversations",
-        ["user_email"],
-        unique=False,
-    )
-    op.create_table(
-        "chat_messages",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("email_conversation_id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "role",
-            sa.Enum("USER", "ASSISTANT", name="messagerole"),
-            nullable=False,
-        ),
-        sa.Column("content", sa.Text(), nullable=False),
-        sa.Column("created_at", sa.TIMESTAMP(timezone=True), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["email_conversation_id"],
-            ["email_conversations.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("chat_messages")
-    op.drop_index(
-        op.f("ix_email_conversations_user_email"),
-        table_name="email_conversations",
-    )
-    op.drop_table("email_conversations")
-    # ### end Alembic commands ###
