--- backend/alembic/versions/bdee7e922ed6_translation_columns_in_articles_and_.py
@@ -1,54 +1,0 @@
-"""Translation columns in Articles and Entities table
-
-Revision ID: bdee7e922ed6
-Revises: 82da1701bd5f
-Create Date: 2025-06-24 03:18:28.626804
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bdee7e922ed6'
-down_revision: Union[str, None] = '82da1701bd5f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('entities',
-    sa.Column('id', sa.Uuid(), nullable=False, server_default=sa.text("'00000000-0000-0000-0000-000000000000'")),
-    sa.Column('entity_type', sa.Text(), nullable=False, server_default=sa.text("'unknown'")),
-    sa.Column('value', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False, server_default=sa.text("'unknown'")),
-    sa.Column('article_id', sa.Uuid(), nullable=False, server_default=sa.text("'00000000-0000-0000-0000-000000000000'")),
-    sa.ForeignKeyConstraint(['article_id'], ['articles.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_entities_article_id'), 'entities', ['article_id'], unique=False)
-    op.add_column('articles', sa.Column('title_en', sa.Text(), nullable=True))
-    op.add_column('articles', sa.Column('title_de', sa.Text(), nullable=True))
-    op.add_column('articles', sa.Column('content_en', sa.Text(), nullable=True))
-    op.add_column('articles', sa.Column('content_de', sa.Text(), nullable=True))
-    op.add_column('articles', sa.Column('summary_en', sa.Text(), nullable=True))
-    op.add_column('articles', sa.Column('summary_de', sa.Text(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('articles', 'summary_de')
-    op.drop_column('articles', 'summary_en')
-    op.drop_column('articles', 'content_de')
-    op.drop_column('articles', 'content_en')
-    op.drop_column('articles', 'title_de')
-    op.drop_column('articles', 'title_en')
-    op.drop_index(op.f('ix_entities_article_id'), table_name='entities')
-    op.drop_table('entities')
-    # ### end Alembic commands ###
