--- backend/alembic/versions/cce014679366_match_topic_id.py
@@ -1,37 +1,0 @@
-"""match topic_id
-
-Revision ID: cce014679366
-Revises: 233ad2724f46
-Create Date: 2025-06-10 18:13:18.721470
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'cce014679366'
-down_revision: Union[str, None] = '233ad2724f46'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('matches', sa.Column('topic_id', sa.Uuid(), nullable=True))
-    op.create_index(op.f('ix_matches_topic_id'), 'matches', ['topic_id'], unique=False)
-    op.create_foreign_key(None, 'matches', 'topics', ['topic_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'matches', type_='foreignkey')
-    op.drop_index(op.f('ix_matches_topic_id'), table_name='matches')
-    op.drop_column('matches', 'topic_id')
-    # ### end Alembic commands ###
