--- backend/alembic/versions/d445a6780cc4_add_username_to_subscriptions.py
@@ -1,41 +1,0 @@
-"""add username to subscriptions
-
-Revision ID: d445a6780cc4
-Revises: f4218eba1395
-Create Date: 2025-06-30 21:28:30.551738
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = "d445a6780cc4"
-down_revision: Union[str, None] = "f4218eba1395"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "subscriptions",
-        sa.Column(
-            "encrypted_username",
-            sa.LargeBinary(),
-            nullable=True,
-        ),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("subscriptions", "encrypted_username")
-    # ### end Alembic commands ###
