--- backend/alembic/versions/d6de83802d3f_change_comment_to_text_add_score_and_.py
@@ -1,58 +1,0 @@
-"""change comment to TEXT, add score and matched_at
-
-Revision ID: d6de83802d3f
-Revises: e2ffa37f4e8f
-Create Date: 2025-06-24 20:40:10.582483
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = "d6de83802d3f"
-down_revision: Union[str, None] = "e2ffa37f4e8f"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("matches", sa.Column("score", sa.Float(), nullable=False, server_default=sa.text('0.0')))
-    op.add_column(
-        "matches",
-        sa.Column(
-            "matched_at",
-            sa.TIMESTAMP(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=False,
-        ),
-    )
-    op.alter_column(
-        "matches",
-        "comment",
-        existing_type=sa.VARCHAR(length=255),
-        type_=sa.Text(),
-        existing_nullable=True,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "matches",
-        "comment",
-        existing_type=sa.Text(),
-        type_=sa.VARCHAR(length=255),
-        existing_nullable=True,
-    )
-    op.drop_column("matches", "matched_at")
-    op.drop_column("matches", "score")
-    # ### end Alembic commands ###
