--- backend/alembic/versions/d8c46ac01508_add_report_model.py
@@ -1,42 +1,0 @@
-"""Add report model
-
-Revision ID: d8c46ac01508
-Revises: 82da1701bd5f
-Create Date: 2025-06-17 17:10:12.809684
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = 'd8c46ac01508'
-down_revision: Union[str, None] = '82da1701bd5f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('reports',
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('search_profile_id', sa.Uuid(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('time_slot', sqlmodel.sql.sqltypes.AutoString(length=32), nullable=True),
-    sa.Column('s3_key', sqlmodel.sql.sqltypes.AutoString(length=512), nullable=False),
-    sa.Column('status', sa.Enum('PENDING', 'UPLOADED', 'FAILED', name='reportstatus'), nullable=False),
-    sa.ForeignKeyConstraint(['search_profile_id'], ['search_profiles.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('reports')
-    # ### end Alembic commands ###
