--- backend/alembic/versions/dcfad6a3ce85_add_translation_columns_for_entities.py
@@ -1,35 +1,0 @@
-"""Add translation columns for entities
-
-Revision ID: dcfad6a3ce85
-Revises: 307c2d987d05
-Create Date: 2025-06-28 13:16:26.465723
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'dcfad6a3ce85'
-down_revision: Union[str, None] = '307c2d987d05'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('entities', sa.Column('value_en', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True))
-    op.add_column('entities', sa.Column('value_de', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('entities', 'value_de')
-    op.drop_column('entities', 'value_en')
-    # ### end Alembic commands ###
