--- backend/alembic/versions/e1e416a56286_add_exteded_rights_to_search_profile.py
@@ -1,35 +1,0 @@
-"""add exteded rights to search profile
-
-Revision ID: e1e416a56286
-Revises: d8c46ac01508
-Create Date: 2025-06-22 17:34:39.602369
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e1e416a56286'
-down_revision: Union[str, None] = 'd8c46ac01508'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('search_profiles', sa.Column('can_read_user_ids', sa.ARRAY(sa.UUID()), server_default='{}', nullable=True))
-    op.add_column('search_profiles', sa.Column('can_edit_user_ids', sa.ARRAY(sa.UUID()), server_default='{}', nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('search_profiles', 'can_edit_user_ids')
-    op.drop_column('search_profiles', 'can_read_user_ids')
-    # ### end Alembic commands ###
